WIP: Move tracers inside readers, as it depends on Victoria. Start decoupling tracing from analysis.
